refactor/TrackComments - 트랙컴포넌트 리펙토링완료 다이나믹 임포트 + 스켈레톤 추가 / 위키api 서버로 관리 + 몽고추가
