+++ src/app/tracks/[id]/page.tsx
@@ -4,7 +4,7 @@ import getTrackIdAlbum from '@/features/tracks/hooks/getTrackIdAlbum';
 import HeaderMain from '@/shared/components/HeaderMain/HeaderMain';
 import TrackDescription from '@/features/tracks/components/TrackDescription/TrackDescription';
 import TrackList from '@/features/tracks/components/TrackList';
-import TrackComments from '@/features/tracks/components/TrackComets/TrackComments';
+import TrackComments from '@/features/tracks/components/TrackCommets/TrackComments';
 
 export const metadata = {
   title: 'Track Page',
