+++ src/features/tracks/components/TrackDescription/TrackDescription.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 
 import useFetchWikiInfo from '@/features/tracks/hooks/TrackDescription/useFetchWikiInfo';
 import { Album } from '@/shared/types/spotifyTrack';
-import TrackCommentsSkeleton from '@/features/tracks/components/TrackComets/TrackCommentsSkeleton';
+import TrackCommentsSkeleton from '@/features/tracks/components/TrackCommets/TrackCommentsSkeleton';
 export default function TrackDescription({ album }: { album: Album }) {
   // 앨범 정보를 위키피디아를 통해 가져옴
   const { summary } = useFetchWikiInfo({ album });
